requirements.txt
# Core dependencies
requests==2.31.0
python-dotenv==1.0.0

# Image processing
Pillow==10.2.0

# WordPress integration
python-wordpress-xmlrpc==2.3

# AI and NLP
anthropic==0.21.3
openai==1.12.0

# Content processing
markdown==3.5.2
beautifulsoup4==4.12.3
html2text==2020.1.16

# Web scraping (for Suno information gathering)
selenium==4.18.1
scrapy==2.11.0

# Development tools
pytest==8.0.0
black==24.1.1
flake8==7.0.0

# Utilities
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2024.1